move some functions of intersect to misc as these functions depend on upper modules
